Combine build scripts for translations
