+++ README.md
@@ -23,7 +23,5 @@ Translation happens by conversion Markdown into _gettext_ using
 [po4a](https://po4a.org). To generate the _.md_ files from the _gettext .po_
 files, run:
 ```bash
-$ python po/filter-markdown.py
-$ po4a po/po4a.conf
-$ python po/set-language.py
+$ python po/build.py
 ```
